+++ server/src/v1/Entry/Service/entry.service.ts
@@ -2,23 +2,37 @@
 import { Injectable, BadRequestException, HttpException, HttpStatus } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 
+// Other dependencies
+import { Validator } from 'class-validator'
+import { ObjectID } from 'mongodb'
+
 // Local files
 import { EntriesRepository } from 'src/shared/Repositories/entries.repository'
 import { EntriesEntity } from 'src/shared/Entities/entries.entity'
 import { CreateEntryDto } from '../Dto/create-entry.dto'
 import { TitlesRepository } from 'src/shared/Repositories/title.repository'
 import { serializerService, ISerializeResponse } from 'src/shared/Services/serializer.service'
+import { UsersRepository } from 'src/shared/Repositories/users.repository'
 
 @Injectable()
 export class EntryService {
+
+    private validator: ObjectID
+
     constructor(
         @InjectRepository(EntriesRepository)
         private readonly entriesRepository: EntriesRepository,
         @InjectRepository(TitlesRepository)
         private readonly titlesRepository: TitlesRepository,
-    ) {}
+        @InjectRepository(UsersRepository)
+        private readonly usersRepository: UsersRepository,
+    ) {
+        this.validator = new Validator()
+    }
 
     async getEntry(entryId: string): Promise<ISerializeResponse> {
+        if (!this.validator.isMongoId(entryId)) throw new BadRequestException(`EntryId must be a MongoId.`)
+
         const entry: EntriesEntity = await this.entriesRepository.getEntry(entryId)
         const id: string = String(entry.id)
         delete entry.id
@@ -31,6 +45,8 @@ export class EntryService {
     }
 
     async updateEntry(updatedBy: string, entryId: string, text: string): Promise<ISerializeResponse> {
+        if (!this.validator.isMongoId(entryId)) throw new BadRequestException('EntryId must be a MongoId.')
+
         const entry: EntriesEntity = await this.entriesRepository.updateEntry(updatedBy, entryId, text)
         const id: string = String(entry.id)
         delete entry.id
@@ -38,6 +54,8 @@ export class EntryService {
     }
 
     async createEntry(writtenBy: string, dto: CreateEntryDto): Promise<HttpException | ISerializeResponse> {
+        if (!this.validator.isMongoId(dto.titleId)) throw new BadRequestException('TitleId must be a MongoId.')
+
         try {
           await this.titlesRepository.updateEntryCount(dto.titleId, true)
         } catch (err) {
@@ -48,10 +66,27 @@ export class EntryService {
         return serializerService.serializeResponse(`entry_detail`, newEntry)
     }
 
+    async upVoteEntry(entryId: string, username: string): Promise<HttpException> {
+        if (!this.validator.isMongoId(entryId)) throw new BadRequestException('EntryId must be a MongoId.')
+
+        await this.usersRepository.addVotedEntry(entryId, username, true)
+        await this.entriesRepository.upVoteEntry(entryId)
+        throw new HttpException('Entry has been up voted.', HttpStatus.OK)
+    }
+
+    async downVoteEntry(entryId: string, username: string): Promise<HttpException> {
+        if (!this.validator.isMongoId(entryId)) throw new BadRequestException('EntryId must be a MongoId.')
+
+        await this.usersRepository.addVotedEntry(entryId, username, false)
+        await this.entriesRepository.downVoteEntry(entryId)
+        throw new HttpException('Entry has been down voted.', HttpStatus.OK)
+    }
+
     async deleteEntry(entryId: string): Promise<HttpException> {
+        if (!this.validator.isMongoId(entryId)) throw new BadRequestException('EntryId must be a MongoId.')
+
         const entry: EntriesEntity = await this.entriesRepository.deleteEntry(entryId)
         await this.titlesRepository.updateEntryCount(entry.title_id, false)
-
         throw new HttpException('Entry has been deleted.', HttpStatus.OK)
     }
 }
